requirements.txt
absl-py==1.1.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.2
Deprecated==1.2.13
Flask==2.1.2
flatbuffers==1.12
gast==0.4.0
gcloud==0.18.3
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.3
grpcio==1.47.0
gunicorn==20.1.0
h5py==3.7.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
jwcrypto==1.3.1
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
numpy==1.23.0
oauth2client==4.1.3
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-jwt==3.3.3
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
typing_extensions==4.3.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
